refactor: merge logics
